+++ cpp/test/GomoryHuTreeTest.cpp
@@ -19,14 +19,14 @@ public:
         ct.addEdge(2, 3, 2);
         ct.build();
 
-        CPPUNIT_ASSERT(ct.edges[0][0] == make_pair(1, 0LL));
-        CPPUNIT_ASSERT(ct.edges[0][1] == make_pair(2, 14LL));
-        CPPUNIT_ASSERT(ct.edges[1][0] == make_pair(0, 0LL));
-        CPPUNIT_ASSERT(ct.edges[2][0] == make_pair(0, 14LL));
-        CPPUNIT_ASSERT(ct.edges[2][1] == make_pair(4, 13LL));
-        CPPUNIT_ASSERT(ct.edges[3][0] == make_pair(4, 2LL));
-        CPPUNIT_ASSERT(ct.edges[4][0] == make_pair(3, 2LL));
-        CPPUNIT_ASSERT(ct.edges[4][1] == make_pair(2, 13LL));
+        CPPUNIT_ASSERT(ct.edges[0][0] == make_pair(1, int64_t(0)));
+        CPPUNIT_ASSERT(ct.edges[0][1] == make_pair(2, int64_t(14)));
+        CPPUNIT_ASSERT(ct.edges[1][0] == make_pair(0, int64_t(0)));
+        CPPUNIT_ASSERT(ct.edges[2][0] == make_pair(0, int64_t(14)));
+        CPPUNIT_ASSERT(ct.edges[2][1] == make_pair(4, int64_t(13)));
+        CPPUNIT_ASSERT(ct.edges[3][0] == make_pair(4, int64_t(2)));
+        CPPUNIT_ASSERT(ct.edges[4][0] == make_pair(3, int64_t(2)));
+        CPPUNIT_ASSERT(ct.edges[4][1] == make_pair(2, int64_t(13)));
     }
 
     CPPUNIT_TEST_SUITE(GomoryHuTreeTest);
