+++ cpp/test/SegmentTreeTest.cpp
@@ -14,11 +14,11 @@ public:
     void test() {
         SegmentTree st(5);
         st.update(1, 1, 2);
-        CPPUNIT_ASSERT_EQUAL(0LL, st.query(3, 5));
-        CPPUNIT_ASSERT_EQUAL(2LL, st.query(1, 5));
+        CPPUNIT_ASSERT_EQUAL(int64_t(0), st.query(3, 5));
+        CPPUNIT_ASSERT_EQUAL(int64_t(2), st.query(1, 5));
         st.update(4, 2, 3);
-        CPPUNIT_ASSERT_EQUAL(9LL, st.query(2, 3));
-        CPPUNIT_ASSERT_EQUAL(10LL, st.query(1, 5));
+        CPPUNIT_ASSERT_EQUAL(int64_t(9), st.query(2, 3));
+        CPPUNIT_ASSERT_EQUAL((int64_t)10, st.query(1, 5));
     }
 
     CPPUNIT_TEST_SUITE(SegmentTreeTest);
